pyproject.toml
[project]
name = "taia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
